libdmabufinfo: Attempt to read real dmabuf size from sysfs
A DMA buffer size calculated from the VMA range (start - end) does
not necessarily correspond to the real size of the buffer. This is
because buffers can be only partially mapped and in some use cases
require the VMA range to exceed the size of the buffer.

Attempt to read the real DMA buffer size from sysfs when available
else fall back to calulating the size from the VMA range.

The real size is also available from fdinfo for buffers with FD
references. The APIs always read FD references before map refernces
so the real buffer size is implicitly used for buffers with any FD
references.

Bug: 234753494
Test: dmabuf_dump
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>
Change-Id: If0259c5b9fc3b47e3a7626fee2b5c5235795aed8