libdmabufinfo: Use ifstream and string to read lines from files.
It is recommended to use ifstream and string instead of FILE or char *
due that the former are RAII (Resource Adquisition Is Initialization)
objects. This means that the ifstream's and string's destructors will be
invoked at the end of the function or scope where they were declared,
releasing resources automatically.

Test: dmabufinfo_test
Bug: 239110298
Change-Id: Ib1b329a86be059745c5f5c8ca1e6778c5d31fc4b